encoder debug checkpoint (added debug suite, clean signal)

+++ include/Encoder.h
@@ -5,7 +5,7 @@ class Encoder {
 public:
     Encoder(int miso, int clk, int cs, int mosi);
     void begin();
-    float readAngle();       // Returns current angle (0-360) and updates rotation count
+    int readAngle();       // Returns current angle (0-360) and updates rotation count
     float getTotalAngle();   // Returns total angle as (rotationCount * 360 + current angle)
 
 private:
